refactor: updateRoom
